requirements.txt
Django==3.0.4
django-filter==2.2.0
djangorestframework==3.11.0
psycopg2==2.8.4
